feat: 方法论扩展与工程优化，深度分析自动挂钩挖掘流水线
- 工程优化：safe_expr 取代裸 eval；回测按日期预切片降为 O(T*N)；fetcher 限流/并发/缓存生效；接入 ruff 与 CI Lint 步骤

- 方法论扩展：method_hub（贝叶斯收缩/TPE/模型平均、Shapley、纳什议价、PCA+有效秩）、twin_model（IRLS 鲁棒预测+鲁棒性曲线、可解释风险模型 MCTR/CCTR/有效下注）

- 因子体系：线性六种（baseline max IC）、显式非线性贪心、GBDT 树、contextual 建模，按样本外增益>10%才升级，否则回退保证可解释性

- 回测输出：统计体检（Newey-West t/bootstrap CI/JB/Ljung-Box/IC 衰减/分位单调性）+ 12 张图表 + LLM 解读（失败降级规则化）

- 深度分析自动挂钩精炼厂（refinery.deep_analysis*），方法学报告新增第十二章；修复净值统计量短样本年化外推导致 Calmar 失真

- 门禁：ruff check src scripts tests 全通过；pytest tests/ 313 passed, 1 skipped

+++ src/agent/graph.py
@@ -21,7 +21,6 @@ from __future__ import annotations
 
 import logging
 import os
-import random
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
@@ -30,13 +29,14 @@ logger = logging.getLogger(__name__)
 import numpy as np
 import pandas as pd
 
-from agent.context import DEFAULT_TURNS as DEFAULT_CONTEXT_TURNS, render_dialogue_context
+from agent.context import DEFAULT_TURNS as DEFAULT_CONTEXT_TURNS
+from agent.context import render_dialogue_context
 from agent.nodes import FactorAgentNodes
 from agent.state import AgentState
 from engine.backtest import FactorBacktester
 from llm.client import LLMClient, load_config
+from rag.learned_library import DEFAULT_LEARNED_PATH, LearnedFactorLibrary
 from rag.retriever import FactorRetriever, rag_vector_enabled
-from rag.learned_library import LearnedFactorLibrary, DEFAULT_LEARNED_PATH
 
 
 class FactorAgent:
@@ -244,7 +244,7 @@ class FactorAgent:
         if kline is None or kline.empty or "date" not in kline.columns:
             return kline, None
         dates = np.sort(pd.to_datetime(kline["date"]).unique())
-        n_test = max(min_test_days, int(round(len(dates) * test_frac)))
+        n_test = max(min_test_days, round(len(dates) * test_frac))
         if n_test >= len(dates) - 1:
             # 数据量不足以切分，退回全量（不报 OOS）
             return kline, None
@@ -380,7 +380,7 @@ class FactorAgent:
                 "ledger_path": summary.get("ledger_path", ""),
                 "warnings": summary.get("warnings", []),
             }
-        except Exception as e:  # noqa: BLE001 —— 影子节点永不中断主流程
+        except Exception as e:
             ft = {"error": f"{type(e).__name__}: {e}", "submitted": 0, "dry_run": 0}
         if ft.get("submitted", 0) > 0:
             head = (
